+++ packages/flp-config-sub-generator/src/app/index.ts
@@ -193,6 +193,8 @@ export default class extends Generator {
         let keysAdded = false;
         if (title) {
             keysAdded = await this._updateI18n(this.manifestPath, this.manifest as Manifest, {
+                semanticObject,
+                action,
                 title,
                 subTitle
             });
@@ -202,8 +204,8 @@ export default class extends Generator {
             {
                 semanticObject,
                 action,
-                title: keysAdded ? `{{${i18nKeyTitle}}}` : title,
-                subTitle: keysAdded && subTitle ? `{{${i18nKeySubTitle}}}` : subTitle
+                title: keysAdded ? `{{${semanticObject}-${action}.${i18nKeyTitle}}}` : title,
+                subTitle: keysAdded && subTitle ? `{{${semanticObject}-${action}.${i18nKeySubTitle}}}` : subTitle
             },
             true,
             this.fs
@@ -218,19 +220,29 @@ export default class extends Generator {
      * @param titles - the titles to be added to the i18n file
      * @param titles.title - the title to be added to the i18n file
      * @param titles.subTitle - the subtitle to be added to the i18n file
+     * @param titles.semanticObject - the semantic object to be used as a prefix to the title and subtitle
+     * @param titles.action - action to be used as a prefix to the title and subtitle
      * @returns true if the i18n file was updated with the key/values
      */
     private async _updateI18n(
         manifestPath: string,
         manifest: Manifest,
-        { title, subTitle }: { title: string; subTitle?: string }
+        {
+            semanticObject,
+            action,
+            title,
+            subTitle
+        }: { semanticObject: string; action: string; title: string; subTitle?: string }
     ): Promise<boolean> {
         let createProps = false;
         const { 'sap.app': i18nPath } = await getI18nPropertiesPaths(manifestPath, manifest);
         try {
-            const i18nEntries = [{ key: i18nKeyTitle, value: title }];
+            const i18nEntries = [{ key: `${semanticObject}-${action}.${i18nKeyTitle}`, value: title }];
             if (subTitle) {
-                i18nEntries.push({ key: i18nKeySubTitle, value: subTitle });
+                i18nEntries.push({
+                    key: `${semanticObject}-${action}.${i18nKeySubTitle}`,
+                    value: subTitle
+                });
             }
             createProps = await createPropertiesI18nEntries(i18nPath, i18nEntries, this.appRootPath, this.fs);
         } catch (error) {
@@ -252,6 +264,7 @@ export default class extends Generator {
         try {
             if (
                 !this.options.launchFlpConfigAsSubGenerator &&
+                this.abort !== true &&
                 isExtensionInstalled(this.vscode, YUI_EXTENSION_ID, YUI_MIN_VER_FILES_GENERATED_MSG)
             ) {
                 this.appWizard?.showInformation(t('info.filesGenerated'), MessageType.notification);