+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/abap-on-btp/questions.test.ts
@@ -36,7 +36,9 @@ jest.mock('../../../../../src/prompts/connectionValidator', () => {
 
 let cfDiscoveredAbapEnvsMock: ServiceInstanceInfo[] = [];
 const uaaCredsMock = {
-    credentials: {}
+    credentials: {
+        uaa: {}
+    }
 };
 jest.mock('@sap/cf-tools', () => {
     return {