+++ packages/cf-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.1.7
+
+### Patch Changes
+
+-   Updated dependencies [ed8a9b9]
+    -   @sap-ux/ui5-config@0.26.2
+    -   @sap-ux/project-access@1.29.6
+
+## 0.1.6
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.5
+
+## 0.1.5
+
+### Patch Changes
+
+-   Updated dependencies [65f15d9]
+    -   @sap-ux/btp-utils@1.0.1
+
+## 0.1.4
+
+### Patch Changes
+
+-   Updated dependencies [29abc73]
+    -   @sap-ux/project-access@1.29.4
+
+## 0.1.3
+
+### Patch Changes
+
+-   Updated dependencies [9980073]
+    -   @sap-ux/btp-utils@1.0.0
+
 ## 0.1.2
 
 ### Patch Changes