+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.7.28
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.6
+-   @sap-ux/telemetry@0.5.60
+
+## 0.7.27
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.5
+-   @sap-ux/telemetry@0.5.59
+
+## 0.7.26
+
+### Patch Changes
+
+-   Updated dependencies [65f15d9]
+    -   @sap-ux/btp-utils@1.0.1
+    -   @sap-ux/telemetry@0.5.58
+
+## 0.7.25
+
+### Patch Changes
+
+-   Updated dependencies [29abc73]
+    -   @sap-ux/project-access@1.29.4
+    -   @sap-ux/telemetry@0.5.57
+
+## 0.7.24
+
+### Patch Changes
+
+-   Updated dependencies [9980073]
+    -   @sap-ux/btp-utils@1.0.0
+    -   @sap-ux/telemetry@0.5.56
+
 ## 0.7.23
 
 ### Patch Changes